team-GenAlpha/poi-scrap/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.1
geopy==2.3.0
overpy==0.7
wikipedia==1.4.0
webdriver-manager==4.0.1
fake-useragent==1.4.0
python-dotenv==1.0.0
json5==0.9.14
lxml==4.9.3